cleanup all mounts/mappings on a broken drive
This should make it easier to cleanly reinstate the drive. I've seen an
intermittent drive failure in production (complete with "IO error" in
the kernel log), which was resolved just by a clean unmount and
reinstatement, but the clean unmount turned out to be unnecessarily
tedious when not performed by the autopilot.