detect drive existence via directory listing only, no stat() calls
This also ensures that when a broken drive is removed, this event is
recognized as such. When a replacement drive is then added at the same
device path, the autopilot will issue the DriveAddedEvent correctly.

All of this without having to do stat() on broken drives, which is
something that I fear might cause the application to hang.