track mounts in both relevant mount namespaces separately
I'm investigating a bizarre case on one of our QA servers where the
autopilot immediately marks all devices as broken on startup, even after
reboot: It opens the LUKS containers, but then mount(1) fails,
complaining that the devices are mounted already. Same problem when I
then remove the broken-flag on the devices.

My suspicion is that the mount in the host mount NS somehow propagates
into the local NS, causing the local mount(1) call to fail as described
above. But even with DEBUG=1, it's difficult to have some visibility
into the differences between mount NS. I don't have a working theory
how, esp. since the problem exists on only one of 50+ more-or-less
identical hosts.

1. With this commit, both relevant mount namespaces (host and local) are
   tracked separately and mount(1) and unmount(1) calls are performed
   only when absolutely necessary, to the best of our knowledge.

2. To reduce the chance of inconsistent state, XFSDevice will not
   consider its mountPath attribute anymore when deciding what to mount
   and unmount. XFSDevice.mountPath is henceforth only used by the
   assignment logic post-setup.

   This change is also necessary to enable change 1: With two separate
   mount namespaces to consider, we would have to track two mountPaths,
   but then we would not be able to satisfy the Device.MountedPath
   interface.